test(discord): split terminal lifecycle coverage
